Add files via upload
This is the tasks from the Python Bootcamp by Jose Portilla. All of this tasks I solve on my onw at first and then I check it with a solutions from the course. Sometimes I also add a solution from the course at the bottom thus you can compare it.